requirements.txt
actions==1.0.3
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
bunch==1.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configobj==5.0.6
configparser==5.0.1
crayons==0.4.0
idna==3.1
importlib-metadata==3.7.0
iniconfig==1.1.1
lxml==4.6.2
more-itertools==8.7.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.2
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2
toml==0.10.2
tqdm==4.58.0
urllib3==1.26.3
wcwidth==0.2.5
webdriver-manager==3.3.0
webdriverdownloader==1.1.0.3
webdrivermanager==0.10.0
zipp==3.4.0
